pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["io_utils"]

[project]
name = "io-utils"
authors = [
    {name = "manyids2", email = "manyids2@github.com"},
]
description = "Utilities."
readme = "README.md"
requires-python = ">=3.8"
keywords = ["io-utils"]
classifiers = [
    "Programming Language :: Python :: 3",
]
dependencies = [
    "pyyaml",
    "pandas",
]
dynamic = ["version"]

